+++ NEWS.rst
@@ -6,7 +6,8 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Minor: Added json::dump_min for dumping a minified json string.
+* Minor: Added json::keys function for getting the keys of an object.
 
 9.0.0
 -----